charuco_calibration: Split utility functions into their own file
